Bocfel: Fix some "comparison always false" warnings

+++ terps/bocfel/zterp.c
@@ -922,7 +922,7 @@ int main(int argc, char **argv)
       die("unknown story type: %s", chunk->name);
     }
 
-    if(chunk->offset > LONG_MAX) die("zcode offset too large");
+    if((uint64_t)chunk->offset > LONG_MAX) die("zcode offset too large");
 
     memory_size = chunk->size;
     story.offset = chunk->offset;
@@ -941,7 +941,7 @@ int main(int argc, char **argv)
   }
 
   if(memory_size < 64) die("story file too small");
-  if(memory_size > SIZE_MAX - 22) die("story file too large");
+  if((uint64_t)memory_size > SIZE_MAX - 22) die("story file too large");
 
   /* It’s possible for a story to be cut short in the middle of an
    * instruction.  If so, the processing loop will run past the end of
